Roughed in javadoc and empty methods for initial FieldSetup object
This class is currently a placeholder and something Mr. B will work on
with the newer members to the programming team. It will encapsulate
field position information at the start of the match and will likely be
useful during auton.

However, do not work on this class without talking to Mr. B first as he
is intending it as a teaching exercise for some of the newer members.